+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- hnefatafl-client: Make the theme configurable.
+- hnefatafl-client: say what address you're connected to..
+- hnefatafl-client: make the theme configurable.
 - Display all users in a game.
 - hnefatafl-client: add sound.
 - hnefatafl-client: make the board size adjustable.